+++ Plugins/Object/Object.hpp
@@ -55,6 +55,7 @@ private:
     ArgumentStack GetInternalObjectType     (ArgumentStack&& args);
     ArgumentStack AcquireItem               (ArgumentStack&& args);
     ArgumentStack SetFacing                 (ArgumentStack&& args);
+    ArgumentStack ClearSpellEffectsOnOthers (ArgumentStack&& args);
 
     CNWSObject *object(ArgumentStack& args);
 };